Replace Explosion component with TimeLimitComponent
