+++ src/components/mod.rs
@@ -8,7 +8,7 @@ mod boss;
 mod consumable;
 mod defense;
 mod enemy;
-mod explosion;
+mod timelimit;
 mod gamemaster;
 mod item;
 mod planet;
@@ -24,7 +24,7 @@ pub use self::{
     consumable::Consumable,
     defense::Defense,
     enemy::{Enemy, EnemySpawnerTag, EnemyType},
-    explosion::Explosion,
+    timelimit::TimeLimitComponent,
     gamemaster::{BossType, GameMaster, Phase, PhaseType},
     item::Item,
     planet::Planet,
